fix(shared): rename LoginProviders.tsx to avoid case collision on Windows
loginProviders.ts and LoginProviders.tsx differ only in case (ignoring
extension). On Windows CI, Vite resolves './LoginProviders' to the .ts
file (case-insensitive FS), causing "LoginProviders is not exported".

Rename to LoginProviderButtons.tsx — barrel re-export unchanged.
